Move Database wrapper to Core
