+++ root/app/Models/Account.php
@@ -15,7 +15,7 @@
 namespace App\Models;
 
 use Exception;
-use App\Models\Database;
+use App\Core\Database;
 use App\Core\ErrorMiddleware;
 
 class Account
